Implement RoadGrid::new with road-centric transition graph
- Renamed `outgoing_roads` to `next_roads` in `RoadGrid` to reflect graph structure.
- Implemented `RoadGrid::new` using a two-pass algorithm:
  - First pass: Quantize endpoints to identify unique nodes and assign `from`/`to` NodeIds to roads.
  - Second pass: Build `next_roads` adjacency map for valid road transitions.
- Updated `RoadId` and `NodeId` to wrap `usize`.
- Added unit tests for graph construction connectivity.
- Cleaned up unused imports in `traffic/src/road.rs`.

+++ traffic/src/cars.rs
@@ -47,7 +47,7 @@ impl CarWorld {
                 let (r, g, b, _a): (u8, u8, u8, u8) = rng.random();  
                 Car::new_on_road(
                     road_grid, 
-            RoadId((x % (road_grid.roads.len() + 1) as i32) as u16),
+            RoadId((x as usize % (road_grid.roads.len() + 1))),
                     Color::from_rgba(r, g, b, 255),
             networks[x as usize].clone(),
             x as u16
